+++ tools/cdb2sockpool/cdb2sockpool.c
@@ -1039,7 +1039,7 @@ static void increase_file_descriptor_limit()
     if (rlp.rlim_cur < rlp.rlim_max) {
         syslog(LOG_INFO, "Increasing soft file descriptor limit from current "
                          "value of %ld to the hard max (%ld).\n",
-               rlp.rlim_cur, rlp.rlim_max);
+               (long int)rlp.rlim_cur, (long int)rlp.rlim_max);
 
         rlp.rlim_cur = rlp.rlim_max;
